python_api/requirements.txt
aniso8601==8.1.1
appdirs==1.4.4
azure-core==1.25.1
azure-storage-blob==12.13.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
cssselect==1.1.0
fake-useragent==0.1.11
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.10
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
msrest==0.7.1
oauthlib==3.2.1
parameterized==0.8.1
parse==1.19.0
pycparser==2.20
pyee==8.1.0
PyJWT==2.0.1
pymongo==4.2.0
pyodbc==4.0.30
pyOpenSSL==20.0.1
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.1
selenium==3.141.0
shortuuid==1.0.9
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.23
tqdm==4.56.0
typing_extensions==4.3.0
urllib3==1.26.3
w3lib==1.22.0
waitress==1.4.4
websockets==8.1
Werkzeug==1.0.1
wfastcgi==3.0.0
wincertstore==0.2
zope.interface==5.2.0
